+++ ecs/column.go
@@ -15,7 +15,7 @@ type column struct {
 }
 
 // newColumn creates a new column for a given type and capacity.
-func newColumn(tp reflect.Type, isRelation bool, target Entity, capacity uint32) column {
+func newColumn(tp reflect.Type, itemSize uintptr, isRelation bool, target Entity, capacity uint32) column {
 	// TODO: should be use a slice instead of an array here?
 	data := reflect.New(reflect.ArrayOf(int(capacity), tp)).Elem()
 	pointer := data.Addr().UnsafePointer()
@@ -23,7 +23,7 @@ func newColumn(tp reflect.Type, isRelation bool, target Entity, capacity uint32)
 	return column{
 		data:       data,
 		pointer:    pointer,
-		itemSize:   sizeOf(tp),
+		itemSize:   itemSize,
 		isRelation: isRelation,
 		target:     target,
 	}